Implemented a dummy getStatus method, now returns random values.
Changed the RequestMethods for the REST API, so the application is now more restful.

Added functionality to set the current execution thread and runnable to null after they are done executing.
Changed the PostAnnounce to annouce the RPi IP Address to the Server Backend. (C#)

+++ RPi-backend/src/main/java/g1/f18/iod/rpi/backend/datastructure/DroneStatus.java
@@ -16,12 +16,12 @@ public class DroneStatus {
     private float heading;
     private boolean armed;
     private boolean armable;
-    private int lastHeartbeat;
+    private long lastHeartbeat;
     private String systemState;
     private String mode;
     
     
-    public DroneStatus(String gps, float battery, float grndSpeed, float heading, boolean armed, int lastHeartbeat, boolean armable, String systemState, String mode){
+    public DroneStatus(String gps, float battery, float grndSpeed, float heading, boolean armed, long lastHeartbeat, boolean armable, String systemState, String mode){
         this.gps = gps;
         this.battery = battery;
         this.grndSpeed = grndSpeed;
@@ -37,7 +37,7 @@ public class DroneStatus {
         return armable;
     }
 
-    public int getLastHeartbeat() {
+    public long getLastHeartbeat() {
         return lastHeartbeat;
     }
